Move report to legacy
