+++ lib/extension/legacy/report.js
@@ -1,9 +1,9 @@
 const zigbeeHerdsmanConverters = require('zigbee-herdsman-converters');
-const logger = require('../util/logger');
-const settings = require('../util/settings');
+const logger = require('../../util/logger');
+const settings = require('../../util/settings');
 const ZNLDP12LM = zigbeeHerdsmanConverters.devices.find((d) => d.model === 'ZNLDP12LM');
-const utils = require('../util/utils');
-const Extension = require('./extension');
+const utils = require('../../util/utils');
+const Extension = require('../extension');
 
 const defaultConfiguration = {
     minimumReportInterval: 3, maximumReportInterval: 300, reportableChange: 1,
